# Speed up query in managed project listing